+++ packages/std/src/testing/mock.rs
@@ -92,9 +92,9 @@ const CANONICAL_LENGTH: usize = 90; // n = 45
 const SHUFFLES_ENCODE: usize = 10;
 const SHUFFLES_DECODE: usize = 2;
 
-// MockPrecompiles zero pads all human addresses to make them fit the canonical_length
+// Mock zero pads all human addresses to make them fit the canonical_length
 // it trims off zeros for the reverse operation.
-// not really smart, but allows us to see a difference (and consistent length for canonical adddresses)
+// not really smart, but allows us to see a difference (and consistent length for canonical addresses)
 #[derive(Copy, Clone)]
 pub struct MockApi {
     /// Length of canonical addresses created with this API. Contracts should not make any assumptions
@@ -227,6 +227,10 @@ impl Api for MockApi {
 
     fn debug(&self, message: &str) {
         println!("{}", message);
+    }
+
+    fn debug_with_gas(&self, message: &str) {
+        println!("{}, gas_left: NA", message);
     }
 }